split eth/op validations
