+++ crates/optimism/consensus/src/lib.rs
@@ -10,12 +10,17 @@
 #![cfg(feature = "optimism")]
 
 use reth_consensus::{Consensus, ConsensusError};
-use reth_consensus_common::validation::{self, validate_header_extradata};
+use reth_consensus_common::validation::{
+    validate_block_pre_execution, validate_header_extradata, validate_header_standalone,
+};
 use reth_primitives::{
     BlockWithSenders, ChainSpec, Header, Receipt, SealedBlock, SealedHeader, EMPTY_OMMER_ROOT_HASH,
     U256,
 };
 use std::{sync::Arc, time::SystemTime};
+
+mod validation;
+pub use validation::validate_block_post_execution;
 
 /// Optimism consensus implementation.
 ///
@@ -40,7 +45,7 @@ impl OptimismBeaconConsensus {
 
 impl Consensus for OptimismBeaconConsensus {
     fn validate_header(&self, header: &SealedHeader) -> Result<(), ConsensusError> {
-        validation::validate_header_standalone(header, &self.chain_spec)?;
+        validate_header_standalone(header, &self.chain_spec)?;
         Ok(())
     }
 
@@ -100,7 +105,7 @@ impl Consensus for OptimismBeaconConsensus {
     }
 
     fn validate_block_pre_execution(&self, block: &SealedBlock) -> Result<(), ConsensusError> {
-        validation::validate_block_pre_execution(block, &self.chain_spec)
+        validate_block_pre_execution(block, &self.chain_spec)
     }
 
     fn validate_block_post_execution(
@@ -108,6 +113,6 @@ impl Consensus for OptimismBeaconConsensus {
         block: &BlockWithSenders,
         receipts: &[Receipt],
     ) -> Result<(), ConsensusError> {
-        validation::validate_block_post_execution(block, &self.chain_spec, receipts)
+        validate_block_post_execution(block, &self.chain_spec, receipts)
     }
 }
